refactor: migrate JavaScript files to TypeScript and update configurations
