+++ website/core/DocsRating.tsx
@@ -16,7 +16,9 @@ const DocsRating = ({label}) => {
 
   const [haveVoted, setHaveVoted] = useState(false);
   const giveFeedback = value => {
+    //  @ts-expect-error
     if (window.ga) {
+      //  @ts-expect-error
       window.ga('send', {
         hitType: 'event',
         eventCategory: 'button',
@@ -37,7 +39,7 @@ const DocsRating = ({label}) => {
           Is this page useful?
           <svg
             className="i_thumbsup"
-            alt="Like"
+            aria-label="Like"
             onClick={() => giveFeedback(1)}
             xmlns="http://www.w3.org/2000/svg"
             viewBox="0 0 81.13 89.76">
@@ -45,7 +47,7 @@ const DocsRating = ({label}) => {
           </svg>
           <svg
             className="i_thumbsdown"
-            alt="Dislike"
+            aria-label="Dislike"
             onClick={() => giveFeedback(0)}
             xmlns="http://www.w3.org/2000/svg"
             viewBox="0 0 81.13 89.76">
