+++ website/package.json
@@ -14,6 +14,7 @@
     "start": "docusaurus start",
     "build": "docusaurus build && yarn run update-redirect ./build/_redirects ./versions.json",
     "build:fast": "PREVIEW_DEPLOY=true yarn run build",
+    "tsc": "npx tsc --noEmit",
     "swizzle": "docusaurus swizzle",
     "deploy": "docusaurus deploy",
     "serve": "docusaurus serve",
@@ -61,7 +62,9 @@
     "sass": "1.76.0"
   },
   "devDependencies": {
-    "@docusaurus/types": "3.6.0",
+    "@docusaurus/module-type-aliases": "^3.6.3",
+    "@docusaurus/tsconfig": "^3.6.3",
+    "@docusaurus/types": "^3.6.3",
     "@react-native-website/lint-examples": "0.0.0",
     "@react-native-website/update-redirects": "0.0.0",
     "alex": "^10.0.0",
@@ -70,6 +73,7 @@
     "glob-promise": "^4.2.2",
     "path": "^0.12.7",
     "remark-cli": "^11.0.0",
-    "remark-lint-no-dead-urls": "^1.1.0"
+    "remark-lint-no-dead-urls": "^1.1.0",
+    "typescript": "^5.7.2"
   }
 }
