+++ website/src/pages/Home/Hero/Devices.tsx
