+++ website/src/pages/Home/Platforms/index.tsx
