+++ website/src/pages/showcase.tsx
@@ -10,6 +10,7 @@ import useBaseUrl from '@docusaurus/useBaseUrl';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import Layout from '@theme/Layout';
 import ThemedImage from '@theme/ThemedImage';
+import users from '../../showcase.json';
 
 import IconExternalLink from '../theme/Icon/ExternalLink';
 
@@ -18,8 +19,13 @@ const renderApp = (app, i) => <AppBox app={app} key={`app-${app.name}-${i}`} />;
 function Section({
   element = 'section',
   children,
-  className,
   background = 'light',
+  className,
+}: {
+  element?: any;
+  children: React.ReactNode;
+  background?: 'light' | 'dark';
+  className?: string;
 }) {
   const El = element;
   return (
@@ -105,9 +111,8 @@ const randomizeApps = apps =>
 
 const Showcase = () => {
   const {siteConfig} = useDocusaurusContext();
-
-  const {meta, microsoft, shopify, wix, amazon, others} =
-    siteConfig.customFields.users;
+  const {meta, microsoft, shopify, wix, amazon, others} = siteConfig
+    .customFields.users as typeof users;
   const [pinnedRandomizedApps, setPinnedRandomizedApps] = useState([]);
   const [randomizedApps, setRandomizedApps] = useState([]);
